Reorganizar classes para v4 conta corrente
