+++ Sicoob.Conta/Models/v4/Transacao.cs
@@ -1,0 +1,8 @@
+using Sicoob.Conta.Models.Shared;
+
+namespace Sicoob.Conta.Models.v4;
+
+public class Transacao : TransacaoBase
+{
+    public string TransactionId { get; set; }
+}
