+++ Sicoob.Conta/SicoobContaCorrenteV4.cs
@@ -2,6 +2,8 @@ using System;
 using System.Net.Http;
 using System.Threading.Tasks;
 using Sicoob.Conta.Models;
+using Sicoob.Conta.Models.Shared;
+using Sicoob.Conta.Models.v2;
 using Sicoob.Shared.Models;
 using Sicoob.Shared.Models.Acesso;
 using Simple.API;
